Moved most of the logic into RadarService
