+++ src/main/java/org/radarcns/android/RadarService.java
@@ -1,13 +1,42 @@
 package org.radarcns.android;
 
+import android.app.AppOpsManager;
+import android.app.Notification;
+import android.app.PendingIntent;
 import android.app.Service;
+import android.bluetooth.BluetoothAdapter;
 import android.content.*;
-import android.os.IBinder;
+import android.content.pm.PackageManager;
+import android.location.LocationManager;
+import android.os.*;
+import android.support.v4.content.ContextCompat;
+import android.widget.Toast;
+import org.radarcns.android.auth.AppAuthState;
+import org.radarcns.android.auth.AppSource;
+import org.radarcns.android.auth.LoginActivity;
+import org.radarcns.android.device.DeviceServiceConnection;
+import org.radarcns.android.device.DeviceServiceProvider;
+import org.radarcns.android.device.DeviceStatusListener;
+import org.radarcns.android.kafka.ServerStatusListener;
+import org.radarcns.android.util.Boast;
+import org.radarcns.data.TimedInt;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.*;
+
+import static android.Manifest.permission.*;
+import static org.radarcns.android.device.DeviceService.DEVICE_CONNECT_FAILED;
+import static org.radarcns.android.device.DeviceService.DEVICE_STATUS_NAME;
+
 
 public abstract class RadarService extends Service {
+    private static final Logger logger = LoggerFactory.getLogger(RadarService.class);
+
     public static String RADAR_PACKAGE = RadarService.class.getPackage().getName();
 
-    public static String EXTRA_CREATOR_CLASS = RADAR_PACKAGE + ".EXTRA_CREATOR_CLASS";
+    public static String EXTRA_MAIN_ACTIVITY = RADAR_PACKAGE + ".EXTRA_MAIN_ACTIVITY";
+    public static String EXTRA_LOGIN_ACTIVITY = RADAR_PACKAGE + ".EXTRA_LOGIN_ACTIVITY";
 
     public static String ACTION_CHECK_PERMISSIONS = RADAR_PACKAGE + ".ACTION_CHECK_PERMISSIONS";
     public static String EXTRA_PERMISSIONS = RADAR_PACKAGE + ".EXTRA_PERMISSIONS";
@@ -29,51 +58,466 @@ public abstract class RadarService extends Service {
         }
     };
 
-    private String creatorClass;
+    private final BroadcastReceiver loginBroadcastReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            updateAuthState(AppAuthState.Builder.from(intent.getExtras()).build());
+        }
+    };
+
+
+    private String mainActivityClass;
+    private String loginActivityClass;
+
+    /** Filters to only listen to certain device IDs. */
+    private final Map<DeviceServiceConnection, Set<String>> deviceFilters = new HashMap<>();
+
+
+    /**
+     * Background handler thread, to do the service orchestration. Having this in the background
+     * is important to avoid any lags in the UI. It is shutdown whenever the activity is not
+     * running.
+     */
+    private HandlerThread mHandlerThread;
+    /** Hander in the background. It is set to null whenever the activity is not running. */
+    private Handler mHandler;
+
+    private boolean isForcedDisconnected;
+
+    /** Defines callbacks for service binding, passed to bindService() */
+    private final BroadcastReceiver bluetoothReceiver;
+    private final BroadcastReceiver deviceFailedReceiver;
+
+    /** Connections. **/
+    private List<DeviceServiceProvider> mConnections;
+
+    /** An overview of how many records have been sent throughout the application. */
+    private final Map<DeviceServiceConnection, TimedInt> mTotalRecordsSent = new HashMap<>();
+    private String latestTopicSent;
+    private final TimedInt latestNumberOfRecordsSent = new TimedInt();
+
+    /** Current server status. */
+    private ServerStatusListener.Status serverStatus;
+    private AppAuthState authState;
+
+    private final LinkedHashSet<String> needsPermissions = new LinkedHashSet<>();
+    private boolean requestedBt;
+
 
     @Override
     public IBinder onBind(Intent intent) {
         return null;
     }
 
+    public RadarService() {
+        super();
+
+
+        bluetoothReceiver = new BroadcastReceiver() {
+            @Override
+            public void onReceive(Context context, Intent intent) {
+                final String action = intent.getAction();
+
+                if (action.equals(BluetoothAdapter.ACTION_STATE_CHANGED)) {
+                    final int state = intent.getIntExtra(BluetoothAdapter.EXTRA_STATE, BluetoothAdapter.ERROR);
+                    logger.info("Bluetooth state {}", state);
+                    // Upon state change, restart ui handler and restart Scanning.
+                    if (state == BluetoothAdapter.STATE_ON) {
+                        logger.info("Bluetooth is on");
+                        startScanning();
+                    } else if (state == BluetoothAdapter.STATE_OFF) {
+                        logger.warn("Bluetooth is off");
+                        startScanning();
+                    }
+                }
+            }
+        };
+
+        deviceFailedReceiver = new BroadcastReceiver() {
+            @Override
+            public void onReceive(Context context, final Intent intent) {
+                if (intent.getAction().equals(DEVICE_CONNECT_FAILED)) {
+
+                    Boast.makeText(RadarService.this, "Cannot connect to device "
+                                    + intent.getStringExtra(DEVICE_STATUS_NAME),
+                            Toast.LENGTH_SHORT).show();
+                }
+            }
+        };
+    }
+
     @Override
     public void onCreate() {
         super.onCreate();
+
+        authState = AppAuthState.Builder.from(this).build();
 
         registerReceiver(configurationBroadcastReceiver,
                 new IntentFilter(RadarConfiguration.RADAR_CONFIGURATION_CHANGED));
         registerReceiver(permissionsBroadcastReceiver,
                 new IntentFilter(ACTION_PERMISSIONS_GRANTED));
+        registerReceiver(loginBroadcastReceiver,
+                new IntentFilter(LoginActivity.ACTION_LOGIN_SUCCESS));
+        registerReceiver(bluetoothReceiver, new IntentFilter(BluetoothAdapter.ACTION_STATE_CHANGED));
+        registerReceiver(deviceFailedReceiver, new IntentFilter(DEVICE_CONNECT_FAILED));
+
 
         configure();
+
+        try {
+            mConnections = DeviceServiceProvider.loadProviders(this, RadarConfiguration.getInstance());
+            // TODO: check what sources are available in the authState (if any)
+            // if any sources are available:
+            //   - only start up providers that DeviceServiceProvider#matches one of the sources
+            //   - set that source in the DeviceServiceProvider.
+            for (DeviceServiceProvider provider : mConnections) {
+                DeviceServiceConnection connection = provider.getConnection();
+                mTotalRecordsSent.put(connection, new TimedInt());
+                deviceFilters.put(connection, Collections.<String>emptySet());
+            }
+        } catch (IllegalArgumentException ex) {
+            logger.error("Cannot instantiate device provider, waiting to fetch to complete", ex);
+            mConnections = Collections.emptyList();
+        }
+
+        checkPermissions();
+
+        new AsyncTask<DeviceServiceProvider, Void, Void>() {
+            @Override
+            protected Void doInBackground(DeviceServiceProvider... params) {
+                for (DeviceServiceProvider provider : params) {
+                    if (!provider.isBound()) {
+                        logger.info("Binding to service: {}", provider);
+                        provider.bind();
+                    } else {
+                        logger.info("Already bound: {}", provider);
+                    }
+                }
+                return null;
+            }
+        }.execute(mConnections.toArray(new DeviceServiceProvider[mConnections.size()]));
+
     }
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
-        creatorClass = intent.getStringExtra(EXTRA_CREATOR_CLASS);
-        return super.onStartCommand(intent, flags, startId);
+        mainActivityClass = intent.getStringExtra(EXTRA_MAIN_ACTIVITY);
+        loginActivityClass = intent.getStringExtra(EXTRA_LOGIN_ACTIVITY);
+
+        startForeground(1,
+                new Notification.Builder(this)
+                        .setContentTitle("RADAR")
+                        .setContentText("Open RADAR app")
+                        .setContentIntent(PendingIntent.getActivity(this, 0, new Intent().setComponent(new ComponentName(this, mainActivityClass)), 0))
+                        .build());
+
+        return START_STICKY;
     }
 
     @Override
     public void onDestroy() {
         unregisterReceiver(configurationBroadcastReceiver);
         unregisterReceiver(permissionsBroadcastReceiver);
+        unregisterReceiver(loginBroadcastReceiver);
+        unregisterReceiver(bluetoothReceiver);
+        unregisterReceiver(deviceFailedReceiver);
+
+        for (DeviceServiceProvider provider : mConnections) {
+            if (provider.isBound()) {
+                logger.info("Unbinding service: {}", provider);
+                provider.unbind();
+            } else {
+                logger.info("Already unbound: {}", provider);
+            }
+        }
 
         super.onDestroy();
     }
 
     protected void configure() {
-        RadarConfiguration configuration = RadarConfiguration.getInstance();
+      //  RadarConfiguration configuration = RadarConfiguration.getInstance();
+
+        for (DeviceServiceProvider provider : mConnections) {
+            provider.updateConfiguration();
+        }
     }
 
     protected void requestPermissions(String[] permissions) {
         startActivity(new Intent()
-                .setComponent(new ComponentName(this, creatorClass))
+                .setComponent(new ComponentName(this, mainActivityClass))
                 .setFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT)
                 .setAction(ACTION_CHECK_PERMISSIONS)
                 .putExtra(EXTRA_PERMISSIONS, permissions));
     }
 
     protected void onPermissionsGranted(String[] permissions, int[] grantResults) {
+        for (int i = 0; i < permissions.length; i++) {
+            if (grantResults[i] == PackageManager.PERMISSION_GRANTED) {
+                logger.info("Granted permission {}", permissions[i]);
+                needsPermissions.remove(permissions[i]);
+            } else {
+                logger.info("Denied permission {}", permissions[i]);
+                return;
+            }
+        }
+        // Permission granted.
+        startScanning();
+    }
+
+    protected void startLogin() {
+        startActivity(new Intent().setComponent(new ComponentName(this, loginActivityClass)));
+    }
+
+    protected void updateAuthState(AppAuthState authState) {
+        this.authState = authState;
+        RadarConfiguration.getInstance().put(RadarConfiguration.PROJECT_ID_KEY, authState.getProjectId());
+        RadarConfiguration.getInstance().put(RadarConfiguration.USER_ID_KEY, authState.getUserId());
+        configure();
+    }
+
+    public void serviceConnected(DeviceServiceConnection<?> connection) {
+        ServerStatusListener.Status status = connection.getServerStatus();
+        logger.info("Initial server status: {}", status);
+        updateServerStatus(status);
+        startScanning();
+    }
+
+    public void serviceDisconnected(final DeviceServiceConnection<?> connection) {
+            new AsyncTask<DeviceServiceConnection, Void, Void>() {
+                @Override
+                protected Void doInBackground(DeviceServiceConnection... params) {
+                    DeviceServiceProvider provider = getConnectionProvider(connection);
+                    logger.info("Rebinding {} after disconnect", provider);
+                    if (provider.isBound()) {
+                        provider.unbind();
+                    }
+                    provider.bind();
+
+                    return null;
+                }
+            }.execute();
+    }
+
+    public void updateServerStatus(ServerStatusListener.Status serverStatus) {
+        this.serverStatus = serverStatus;
+        if (serverStatus == ServerStatusListener.Status.UNAUTHORIZED) {
+            synchronized (RadarService.this) {
+                // login already started, or was finished up to 3 seconds ago (give time to propagate new auth state.)
+                if (authState.isInvalidated() || authState.timeSinceLastUpdate() < 3_000L) {
+                    return;
+                }
+                authState.invalidate(RadarService.this);
+            }
+            startLogin();
+        }
+    }
+
+    public void deviceStatusUpdated(final DeviceServiceConnection<?> connection, final DeviceStatusListener.Status status) {
+        new Handler(Looper.getMainLooper()).post(new Runnable() {
+            @Override
+            public void run() {
+                int showRes = -1;
+                switch (status) {
+                    case READY:
+                        showRes = R.string.device_ready;
+                        break;
+                    case CONNECTED:
+                        showRes = R.string.device_connected;
+                        break;
+                    case CONNECTING:
+                        showRes = R.string.device_connecting;
+                        logger.info( "Device name is {} while connecting.", connection.getDeviceName());
+                        break;
+                    case DISCONNECTED:
+                        showRes = R.string.device_disconnected;
+                        startScanning();
+                        break;
+                    default:
+                        break;
+                }
+                if (showRes != -1) {
+                    Boast.makeText(RadarService.this, showRes).show();
+                }
+            }
+        });
+    }
+
+    public void updateServerRecordsSent(DeviceServiceConnection<?> connection, String topic, int numberOfRecords) {
+        if (numberOfRecords >= 0){
+            mTotalRecordsSent.get(connection).add(numberOfRecords);
+        }
+        latestTopicSent = topic;
+        latestNumberOfRecordsSent.set(numberOfRecords);
+    }
+
+    protected DeviceServiceProvider getConnectionProvider(DeviceServiceConnection<?> connection) {
+        for (DeviceServiceProvider provider : mConnections) {
+            if (provider.getConnection().equals(connection)) {
+                return provider;
+            }
+        }
+        throw new IllegalArgumentException("DeviceServiceConnection "
+                + connection.getServiceClassName() + " not set");
+    }
+
+
+    protected void startScanning() {
+        if (isForcedDisconnected) {
+            return;
+        }
+        requestedBt = false;
+        for (DeviceServiceProvider<?> provider : mConnections) {
+            DeviceServiceConnection connection = provider.getConnection();
+            if (!connection.hasService() || connection.isRecording() || !checkPermissions(provider)) {
+                continue;
+            }
+
+            logger.info("Starting recording on connection {}", connection);
+            AppSource source = provider.getSource();
+            Set<String> filters;
+            if (source != null && source.getExpectedSourceName() != null) {
+                String[] expectedIds = source.getExpectedSourceName().split(",");
+                filters = new HashSet<>(Arrays.asList(expectedIds));
+            } else {
+                filters = deviceFilters.get(connection);
+            }
+            connection.startRecording(filters);
+        }
+    }
+
+    protected void checkPermissions() {
+        List<String> permissions = new ArrayList<>();
+        permissions.addAll(getServicePermissions());
+        for (DeviceServiceProvider<?> provider : mConnections) {
+            permissions.addAll(provider.needsPermissions());
+        }
+
+        needsPermissions.clear();
+
+        for (String permission : permissions) {
+            if (permission.equals(ACCESS_FINE_LOCATION) || permission.equals(ACCESS_COARSE_LOCATION)) {
+                LocationManager locationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
+                boolean isGpsEnabled = locationManager.isProviderEnabled(LocationManager.GPS_PROVIDER);
+                boolean isNetworkEnabled = locationManager.isProviderEnabled(LocationManager.NETWORK_PROVIDER);
+
+                //Start your Activity if location was enabled:
+                if (!isGpsEnabled && !isNetworkEnabled) {
+                    needsPermissions.add(LOCATION_SERVICE);
+                    needsPermissions.add(permission);
+                }
+            }
+
+            if (permission.equals(PACKAGE_USAGE_STATS)) {
+                AppOpsManager appOps = (AppOpsManager) getSystemService(Context.APP_OPS_SERVICE);
+                int mode = appOps.checkOpNoThrow(
+                        "android:get_usage_stats", android.os.Process.myUid(), getPackageName());
+
+                if (mode != AppOpsManager.MODE_ALLOWED) {
+                    needsPermissions.add(permission);
+                }
+            } else if (ContextCompat.checkSelfPermission(this, permission) != PackageManager
+                    .PERMISSION_GRANTED) {
+                logger.info("Need to request permission for {}", permission);
+                needsPermissions.add(permission);
+            }
+        }
+
+        requestPermissions(needsPermissions.toArray(new String[needsPermissions.size()]));
+    }
+
+    protected List<String> getServicePermissions() {
+        return Arrays.asList(ACCESS_NETWORK_STATE, INTERNET);
+    }
+
+
+    protected boolean checkPermissions(DeviceServiceProvider<?> provider) {
+        List<String> providerPermissions = provider.needsPermissions();
+
+        if (providerPermissions.contains(BLUETOOTH)) {
+            if (requestedBt || requestEnableBt()) {
+                logger.info("Cannot start scanning on service {} until bluetooth is turned on.",
+                        provider.getConnection());
+                requestedBt = true;
+                return false;
+            }
+        }
+        for (String permission : providerPermissions) {
+            if (needsPermissions.contains(permission)) {
+                // cannot start
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * Sends an intent to request bluetooth to be turned on.
+     * @return whether a request was sent
+     */
+    protected boolean requestEnableBt() {
+        BluetoothAdapter btAdaptor = BluetoothAdapter.getDefaultAdapter();
+        if (!btAdaptor.isEnabled()) {
+            Intent btIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
+            btIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            getApplicationContext().startActivity(btIntent);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    /** Disconnect from all services. */
+    protected void disconnect() {
+        for (DeviceServiceProvider provider : mConnections) {
+            disconnect(provider.getConnection());
+        }
+    }
+
+    /** Disconnect from given service. */
+    public void disconnect(DeviceServiceConnection connection) {
+        if (connection.isRecording()) {
+            connection.stopRecording();
+        }
+    }
+
+    public ServerStatusListener.Status getServerStatus() {
+        return serverStatus;
+    }
+
+    public TimedInt getTopicsSent(DeviceServiceConnection connection) {
+        return mTotalRecordsSent.get(connection);
+    }
+
+    public String getLatestTopicSent() {
+        return latestTopicSent;
+    }
+
+    public TimedInt getLatestNumberOfRecordsSent() {
+        return latestNumberOfRecordsSent;
+    }
+
+    public List<DeviceServiceProvider> getConnections() {
+        return Collections.unmodifiableList(mConnections);
+    }
+
+    public AppAuthState getAuthState() {
+        return authState;
+    }
+
+    public void setAllowedDeviceIds(final DeviceServiceConnection connection, Set<String> allowedIds) {
+        deviceFilters.put(connection, allowedIds);
+
+        // Do NOT disconnect if input has not changed, is empty or equals the connected device.
+        if (connection.hasService() && !connection.isAllowedDevice(allowedIds)) {
+            new Handler(Looper.getMainLooper()).post(new Runnable() {
+                @Override
+                public void run() {
+                    if (connection.isRecording()) {
+                        connection.stopRecording();
+                        // will restart recording once the status is set to disconnected.
+                    }
+                }
+            });
+        }
     }
 }
