+++ src/main/java/org/radarcns/android/device/DeviceServiceConnection.java
@@ -24,7 +24,7 @@ import android.content.IntentFilter;
 import android.os.IBinder;
 import android.support.annotation.NonNull;
 
-import org.radarcns.android.MainActivity;
+import org.radarcns.android.RadarService;
 import org.radarcns.android.kafka.ServerStatusListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,7 +38,7 @@ import static org.radarcns.android.device.DeviceService.SERVER_STATUS_CHANGED;
 
 public class DeviceServiceConnection<S extends BaseDeviceState> extends BaseServiceConnection<S> {
     private static final Logger logger = LoggerFactory.getLogger(DeviceServiceConnection.class);
-    private final MainActivity mainActivity;
+    private final RadarService radarService;
 
     private final BroadcastReceiver statusReceiver = new BroadcastReceiver() {
         @Override
@@ -51,7 +51,7 @@ public class DeviceServiceConnection<S extends BaseDeviceState> extends BaseServ
                     }
                     setDeviceStatus(DeviceStatusListener.Status.values()[intent.getIntExtra(DEVICE_STATUS_CHANGED, 0)]);
                     logger.info("Updated device status to {}", getDeviceStatus());
-                    mainActivity.deviceStatusUpdated(DeviceServiceConnection.this, getDeviceStatus());
+                    radarService.deviceStatusUpdated(DeviceServiceConnection.this, getDeviceStatus());
                 }
             }
         }
@@ -62,11 +62,11 @@ public class DeviceServiceConnection<S extends BaseDeviceState> extends BaseServ
         public void onReceive(Context context, Intent intent) {
             if (intentMatches(intent, SERVER_STATUS_CHANGED)) {
                 final ServerStatusListener.Status status = ServerStatusListener.Status.values()[intent.getIntExtra(SERVER_STATUS_CHANGED, 0)];
-                mainActivity.updateServerStatus(status);
+                radarService.updateServerStatus(status);
             } else if (intentMatches(intent, SERVER_RECORDS_SENT_TOPIC)) {
                 String topic = intent.getStringExtra(SERVER_RECORDS_SENT_TOPIC); // topicName that updated
                 int numberOfRecordsSent = intent.getIntExtra(SERVER_RECORDS_SENT_NUMBER, 0);
-                mainActivity.updateServerRecordsSent(DeviceServiceConnection.this, topic, numberOfRecordsSent);
+                radarService.updateServerRecordsSent(DeviceServiceConnection.this, topic, numberOfRecordsSent);
             }
         }
     };
@@ -76,25 +76,25 @@ public class DeviceServiceConnection<S extends BaseDeviceState> extends BaseServ
                 && getServiceClassName().equals(intent.getStringExtra(DEVICE_SERVICE_CLASS));
     }
 
-    public DeviceServiceConnection(@NonNull MainActivity mainActivity, String serviceClassName) {
+    public DeviceServiceConnection(@NonNull RadarService radarService, String serviceClassName) {
         super(serviceClassName);
-        this.mainActivity = mainActivity;
+        this.radarService = radarService;
     }
 
     @Override
     public void onServiceConnected(final ComponentName className,
                                    IBinder service) {
-        mainActivity.registerReceiver(statusReceiver,
+        radarService.registerReceiver(statusReceiver,
                 new IntentFilter(DEVICE_STATUS_CHANGED));
 
         IntentFilter serverStatusFilter = new IntentFilter();
         serverStatusFilter.addAction(SERVER_STATUS_CHANGED);
         serverStatusFilter.addAction(SERVER_RECORDS_SENT_TOPIC);
-        mainActivity.registerReceiver(serverStatusListener, serverStatusFilter);
+        radarService.registerReceiver(serverStatusListener, serverStatusFilter);
 
         if (!hasService()) {
             super.onServiceConnected(className, service);
-            mainActivity.serviceConnected(this);
+            radarService.serviceConnected(this);
         }
     }
 
@@ -104,9 +104,9 @@ public class DeviceServiceConnection<S extends BaseDeviceState> extends BaseServ
         super.onServiceDisconnected(className);
 
         if (hadService) {
-            mainActivity.unregisterReceiver(statusReceiver);
-            mainActivity.unregisterReceiver(serverStatusListener);
-            mainActivity.serviceDisconnected(this);
+            radarService.unregisterReceiver(statusReceiver);
+            radarService.unregisterReceiver(serverStatusListener);
+            radarService.serviceDisconnected(this);
         }
     }
 }
